+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added the processor `:securepay` to `Minfraud::Components::Payment`.
 * Ruby 3.2+ is now required. If you're using Ruby 3.0 or 3.1, please use
   version 2.8.0 of this gem.
+* Added `/event/type` values `:credit_application` and `:fund_transfer` to
+  `Minfraud::Components::Event`.
 
 ## v2.8.0 (2025-05-23)